:card_file_box: add: 컴포넌트 폴더 분리
