+++ src/App.js
@@ -1,8 +1,8 @@
 import { useState, useEffect } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { GlobalStyle } from "./GlobalStyle";
-import PostWithImg from "./components/common/PostWithImg";
-import PostOnlyText from "./components/common/PostOnlyText";
+import PostWithImg from "./components/common/PostFormat/PostWithImg/PostWithImg";
+import PostOnlyText from "./components/common/PostFormat/PostOnlyText/PostOnlyText";
 import Navbar from "./components/Navbar/Navbar";
 import Profile from "./pages/Profile/Profile";
 import PostDetail from "./components/Posts/PostDetail";
@@ -99,10 +99,10 @@ function App() {
               <Route path="/profile" element={<Profile />} />
               <Route path="/post" element={<Post />} />
             </Routes>
-            {/* 
+
             <PostOnlyText />
             <PostWithImg />
-            <PostDetail /> */}
+            <PostDetail />
           </Main>
           <Navbar />
         </BrowserRouter>
