+++ src/components/Posts/Comment/CommentAdd/styledCommentAdd.jsx
@@ -1,0 +1,25 @@
+import styled from "styled-components";
+
+export const CommentAddWrapper = styled.section`
+  display: flex;
+  justify-content: space-between;
+  margin: 0 auto;
+  padding: 13px 0px 12px 0px;
+  border-top: 0.5px solid #dbdbdb;
+  bottom: 0;
+  img {
+    width: 36px;
+    height: 36px;
+  }
+  input {
+    flex-grow: 1;
+    margin-left: 18px;
+    margin-right: 18px;
+    border: none;
+  }
+  button {
+    color: #c4c4c4;
+    font-weight: 500;
+    line-height: 17.53px;
+  }
+`;
