+++ src/components/Posts/PostBtn/PostCommentBtn/styledPostCommentBtn.jsx
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const CommentWrapper = styled.button`
+  color: #757575;
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 12px;
+  display: flex;
+  align-items: center;
+  img {
+    width: 20px;
+    height: 20px;
+    margin-right: 6px;
+  }
+`;
