+++ src/components/Posts/PostBtn/PostHearBtn/styledPostHeartBtn.jsx
@@ -1,0 +1,18 @@
+import styled from "styled-components";
+
+export const HeartWrapper = styled.button`
+  color: #757575;
+  font-size: 12px;
+  font-weight: 400;
+  line-height: 12px;
+  display: flex;
+  align-items: center;
+  img {
+    width: 20px;
+    height: 20px;
+    margin-right: 6px;
+  }
+  span {
+    margin-right: 16px;
+  }
+`;
