+++ src/components/common/Modal/Alert/styledDeleteAlert.jsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
 
-const DeleteAlertWrapper = styled.section`
+export const DeleteAlertWrapper = styled.section`
   position: fixed;
   top: 0;
   left: 0;
@@ -46,25 +46,3 @@ const DeleteAlertWrapper = styled.section`
     color: red;
   }
 `;
-
-const DeleteAlert = () => {
-  return (
-    <>
-      <DeleteAlertWrapper>
-        <h3 className="hidden">경고 알림 창</h3>
-        <div>
-          <strong>게시글을 삭제할까요?</strong>
-          <ul>
-            <li>
-              <button>취소</button>
-            </li>
-            <li>
-              <button className="alert-del">삭제</button>
-            </li>
-          </ul>
-        </div>
-      </DeleteAlertWrapper>
-    </>
-  );
-};
-export default DeleteAlert;
