+++ src/components/common/Modal/Modal/styledModal.jsx
@@ -1,9 +1,6 @@
-// 내가 작성한 댓글 : 삭제
-// 다른 사용자가 작성한 댓글 : 신고하기
-
 import styled from "styled-components";
 
-const ModalWrapper = styled.section`
+export const ModalWrapper = styled.section`
   /* 모달창 외부 */
   position: fixed;
   top: 0;
@@ -58,26 +55,3 @@ const ModalWrapper = styled.section`
     height: 100%;
   }
 `;
-// 내가 작성한 댓글 : 삭제
-// 다른 사용자가 작성한 댓글 : 신고하기
-
-// 내가 작성한 게시글 : 삭제, 수정
-// 다른 사용자가 작성한 게시글 : 신고하기
-const Modal = () => {
-  return (
-    <>
-      <ModalWrapper>
-        <h2 className="hidden">댓글 모달창</h2>
-        <ul>
-          <li>
-            <button>삭제</button>
-          </li>
-          {/* <li>
-            <button>신고하기</button>
-          </li> */}
-        </ul>
-      </ModalWrapper>
-    </>
-  );
-};
-export default Modal;
